# internal/ogen/api/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package api

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

var (
	rn12AllowedHeaders = map[string]string{
		"GET": "Authorization,X-Api-Key-L7nqe",
	}
	rn5AllowedHeaders = map[string]string{
		"GET": "X-Api-Key-L7nqe",
	}
	rn7AllowedHeaders = map[string]string{
		"GET": "X-Api-Key-L7nqe",
	}
	rn8AllowedHeaders = map[string]string{
		"GET": "Authorization",
	}
	rn10AllowedHeaders = map[string]string{
		"GET": "Authorization",
	}
	rn13AllowedHeaders = map[string]string{
		"GET": "Authorization",
	}
	rn17AllowedHeaders = map[string]string{
		"GET": "Authorization",
	}
	rn15AllowedHeaders = map[string]string{
		"GET": "Authorization",
	}
	rn18AllowedHeaders = map[string]string{
		"GET": "Authorization",
	}
	rn1AllowedHeaders = map[string]string{
		"GET": "Authorization,X-Api-Key-L7nqe",
	}
	rn20AllowedHeaders = map[string]string{
		"GET": "Authorization",
	}
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"

			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'a': // Prefix: "a"

				if l := len("a"); len(elem) >= l && elem[0:l] == "a" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'n': // Prefix: "nd-orf7kpd"

					if l := len("nd-orf7kpd"); len(elem) >= l && elem[0:l] == "nd-orf7kpd" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleComplexAuthenticationG9wTcRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "GET",
								allowedHeaders: rn12AllowedHeaders,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}

				case 'p': // Prefix: "pi-key-"

					if l := len("pi-key-"); len(elem) >= l && elem[0:l] == "pi-key-" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'c': // Prefix: "cookieh8ryu"

						if l := len("cookieh8ryu"); len(elem) >= l && elem[0:l] == "cookieh8ryu" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "GET":
								s.handleApiKeyCookieD4tZuRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "GET",
									allowedHeaders: nil,
									acceptPost:     "",
									acceptPatch:    "",
								})
							}

							return
						}

					case 'h': // Prefix: "headerx3nws"

						if l := len("headerx3nws"); len(elem) >= l && elem[0:l] == "headerx3nws" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "GET":
								s.handleApiKeyHeaderL5mQaRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "GET",
									allowedHeaders: rn5AllowedHeaders,
									acceptPost:     "",
									acceptPatch:    "",
								})
							}

							return
						}

					case 'q': // Prefix: "queryc6jfa"

						if l := len("queryc6jfa"); len(elem) >= l && elem[0:l] == "queryc6jfa" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "GET":
								s.handleApiKeyQueryW7pErRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "GET",
									allowedHeaders: nil,
									acceptPost:     "",
									acceptPatch:    "",
								})
							}

							return
						}

					case 'w': // Prefix: "with-rolek2vdp"

						if l := len("with-rolek2vdp"); len(elem) >= l && elem[0:l] == "with-rolek2vdp" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "GET":
								s.handleApiKeyWithRoleP9bLcRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "GET",
									allowedHeaders: rn7AllowedHeaders,
									acceptPost:     "",
									acceptPatch:    "",
								})
							}

							return
						}

					}

				}

			case 'b': // Prefix: "b"

				if l := len("b"); len(elem) >= l && elem[0:l] == "b" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'a': // Prefix: "asicn5sge"

					if l := len("asicn5sge"); len(elem) >= l && elem[0:l] == "asicn5sge" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleBasicAuthenticationE6qRtRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "GET",
								allowedHeaders: rn8AllowedHeaders,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}

				case 'e': // Prefix: "earerw7cqm"

					if l := len("earerw7cqm"); len(elem) >= l && elem[0:l] == "earerw7cqm" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleBearerAuthenticationT3nVkRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "GET",
								allowedHeaders: rn10AllowedHeaders,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}

				}

			case 'i': // Prefix: "inherited-securitya8k2z"

				if l := len("inherited-securitya8k2z"); len(elem) >= l && elem[0:l] == "inherited-securitya8k2z" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleInheritedSecurityF3vNxRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET",
							allowedHeaders: rn13AllowedHeaders,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}

			case 'o': // Prefix: "o"

				if l := len("o"); len(elem) >= l && elem[0:l] == "o" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'a': // Prefix: "auth2"

					if l := len("auth2"); len(elem) >= l && elem[0:l] == "auth2" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case '-': // Prefix: "-without-scopesd3wnf"

						if l := len("-without-scopesd3wnf"); len(elem) >= l && elem[0:l] == "-without-scopesd3wnf" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "GET":
								s.handleOauth2WithoutScopesB7vQeRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "GET",
									allowedHeaders: rn17AllowedHeaders,
									acceptPost:     "",
									acceptPatch:    "",
								})
							}

							return
						}

					case 'z': // Prefix: "z6pua"

						if l := len("z6pua"); len(elem) >= l && elem[0:l] == "z6pua" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "GET":
								s.handleOauth2AuthenticationM2cFwRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "GET",
									allowedHeaders: rn15AllowedHeaders,
									acceptPost:     "",
									acceptPatch:    "",
								})
							}

							return
						}

					}

				case 'p': // Prefix: "ptional-authenticationp9tql"

					if l := len("ptional-authenticationp9tql"); len(elem) >= l && elem[0:l] == "ptional-authenticationp9tql" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleOptionalAuthenticationC2yHsRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "GET",
								allowedHeaders: rn18AllowedHeaders,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}

				case 'r': // Prefix: "rq5lzc"

					if l := len("rq5lzc"); len(elem) >= l && elem[0:l] == "rq5lzc" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleAlternativeAuthenticationV8xNpRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "GET",
								allowedHeaders: rn1AllowedHeaders,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}

				}

			case 'p': // Prefix: "publicm4vrb"

				if l := len("publicm4vrb"); len(elem) >= l && elem[0:l] == "publicm4vrb" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handlePublicEndpointR8kWdRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET",
							allowedHeaders: nil,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}

			case 'r': // Prefix: "referenced-security-schemer4cnj"

				if l := len("referenced-security-schemer4cnj"); len(elem) >= l && elem[0:l] == "referenced-security-schemer4cnj" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleReferencedSecuritySchemeU2kMfRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET",
							allowedHeaders: rn20AllowedHeaders,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}

			}

		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name           string
	summary        string
	operationID    string
	operationGroup string
	pathPattern    string
	count          int
	args           [0]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// OperationGroup returns the x-ogen-operation-group value.
func (r Route) OperationGroup() string {
	return r.operationGroup
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"

			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'a': // Prefix: "a"

				if l := len("a"); len(elem) >= l && elem[0:l] == "a" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'n': // Prefix: "nd-orf7kpd"

					if l := len("nd-orf7kpd"); len(elem) >= l && elem[0:l] == "nd-orf7kpd" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "GET":
							r.name = ComplexAuthenticationG9wTcOperation
							r.summary = ""
							r.operationID = "complexAuthenticationG9wTc"
							r.operationGroup = ""
							r.pathPattern = "/and-orf7kpd"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				case 'p': // Prefix: "pi-key-"

					if l := len("pi-key-"); len(elem) >= l && elem[0:l] == "pi-key-" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'c': // Prefix: "cookieh8ryu"

						if l := len("cookieh8ryu"); len(elem) >= l && elem[0:l] == "cookieh8ryu" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "GET":
								r.name = ApiKeyCookieD4tZuOperation
								r.summary = ""
								r.operationID = "apiKeyCookieD4tZu"
								r.operationGroup = ""
								r.pathPattern = "/api-key-cookieh8ryu"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}

					case 'h': // Prefix: "headerx3nws"

						if l := len("headerx3nws"); len(elem) >= l && elem[0:l] == "headerx3nws" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "GET":
								r.name = ApiKeyHeaderL5mQaOperation
								r.summary = ""
								r.operationID = "apiKeyHeaderL5mQa"
								r.operationGroup = ""
								r.pathPattern = "/api-key-headerx3nws"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}

					case 'q': // Prefix: "queryc6jfa"

						if l := len("queryc6jfa"); len(elem) >= l && elem[0:l] == "queryc6jfa" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "GET":
								r.name = ApiKeyQueryW7pErOperation
								r.summary = ""
								r.operationID = "apiKeyQueryW7pEr"
								r.operationGroup = ""
								r.pathPattern = "/api-key-queryc6jfa"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}

					case 'w': // Prefix: "with-rolek2vdp"

						if l := len("with-rolek2vdp"); len(elem) >= l && elem[0:l] == "with-rolek2vdp" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "GET":
								r.name = ApiKeyWithRoleP9bLcOperation
								r.summary = ""
								r.operationID = "apiKeyWithRoleP9bLc"
								r.operationGroup = ""
								r.pathPattern = "/api-key-with-rolek2vdp"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}

					}

				}

			case 'b': // Prefix: "b"

				if l := len("b"); len(elem) >= l && elem[0:l] == "b" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'a': // Prefix: "asicn5sge"

					if l := len("asicn5sge"); len(elem) >= l && elem[0:l] == "asicn5sge" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "GET":
							r.name = BasicAuthenticationE6qRtOperation
							r.summary = ""
							r.operationID = "basicAuthenticationE6qRt"
							r.operationGroup = ""
							r.pathPattern = "/basicn5sge"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				case 'e': // Prefix: "earerw7cqm"

					if l := len("earerw7cqm"); len(elem) >= l && elem[0:l] == "earerw7cqm" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "GET":
							r.name = BearerAuthenticationT3nVkOperation
							r.summary = ""
							r.operationID = "bearerAuthenticationT3nVk"
							r.operationGroup = ""
							r.pathPattern = "/bearerw7cqm"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				}

			case 'i': // Prefix: "inherited-securitya8k2z"

				if l := len("inherited-securitya8k2z"); len(elem) >= l && elem[0:l] == "inherited-securitya8k2z" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "GET":
						r.name = InheritedSecurityF3vNxOperation
						r.summary = ""
						r.operationID = "inheritedSecurityF3vNx"
						r.operationGroup = ""
						r.pathPattern = "/inherited-securitya8k2z"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			case 'o': // Prefix: "o"

				if l := len("o"); len(elem) >= l && elem[0:l] == "o" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'a': // Prefix: "auth2"

					if l := len("auth2"); len(elem) >= l && elem[0:l] == "auth2" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case '-': // Prefix: "-without-scopesd3wnf"

						if l := len("-without-scopesd3wnf"); len(elem) >= l && elem[0:l] == "-without-scopesd3wnf" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "GET":
								r.name = Oauth2WithoutScopesB7vQeOperation
								r.summary = ""
								r.operationID = "oauth2WithoutScopesB7vQe"
								r.operationGroup = ""
								r.pathPattern = "/oauth2-without-scopesd3wnf"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}

					case 'z': // Prefix: "z6pua"

						if l := len("z6pua"); len(elem) >= l && elem[0:l] == "z6pua" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "GET":
								r.name = Oauth2AuthenticationM2cFwOperation
								r.summary = ""
								r.operationID = "oauth2AuthenticationM2cFw"
								r.operationGroup = ""
								r.pathPattern = "/oauth2z6pua"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}

					}

				case 'p': // Prefix: "ptional-authenticationp9tql"

					if l := len("ptional-authenticationp9tql"); len(elem) >= l && elem[0:l] == "ptional-authenticationp9tql" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "GET":
							r.name = OptionalAuthenticationC2yHsOperation
							r.summary = ""
							r.operationID = "optionalAuthenticationC2yHs"
							r.operationGroup = ""
							r.pathPattern = "/optional-authenticationp9tql"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				case 'r': // Prefix: "rq5lzc"

					if l := len("rq5lzc"); len(elem) >= l && elem[0:l] == "rq5lzc" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "GET":
							r.name = AlternativeAuthenticationV8xNpOperation
							r.summary = ""
							r.operationID = "alternativeAuthenticationV8xNp"
							r.operationGroup = ""
							r.pathPattern = "/orq5lzc"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				}

			case 'p': // Prefix: "publicm4vrb"

				if l := len("publicm4vrb"); len(elem) >= l && elem[0:l] == "publicm4vrb" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "GET":
						r.name = PublicEndpointR8kWdOperation
						r.summary = ""
						r.operationID = "publicEndpointR8kWd"
						r.operationGroup = ""
						r.pathPattern = "/publicm4vrb"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			case 'r': // Prefix: "referenced-security-schemer4cnj"

				if l := len("referenced-security-schemer4cnj"); len(elem) >= l && elem[0:l] == "referenced-security-schemer4cnj" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "GET":
						r.name = ReferencedSecuritySchemeU2kMfOperation
						r.summary = ""
						r.operationID = "referencedSecuritySchemeU2kMf"
						r.operationGroup = ""
						r.pathPattern = "/referenced-security-schemer4cnj"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			}

		}
	}
	return r, false
}